+++ packages/dumb/src/jest.config.js
@@ -29,8 +29,6 @@
  * https://jestjs.io/docs/configuration
  */
 
-const esModulesToExclude = ['d3', 'd3-array'].join('|');
-
 module.exports = {
   // All imported modules in your tests should be mocked automatically
   // automock: false,
@@ -112,8 +110,9 @@ module.exports = {
 
   // A map from regular expressions to module names or to arrays of module names that allow to stub out resources with a single module
   moduleNameMapper: {
-    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>../../../__mocks__/fileMock.js",
-    "\\.(css|less)$": "<rootDir>../../../__mocks__/styleMock.js"
+    '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': '<rootDir>../../../__mocks__/fileMock.js',
+    '\\.(css|less)$': '<rootDir>../../../__mocks__/styleMock.js',
+    '^d3-(.*)$': '<rootDir>../../../node_modules/d3-$1/dist/d3-$1'
   },
 
   // An array of regexp pattern strings, matched against all module paths before considered 'visible' to the module loader
@@ -180,7 +179,7 @@ module.exports = {
 
   // The glob patterns Jest uses to detect test files
   testMatch: [
-    '<rootDir>/**/*.test.ts*',
+    '<rootDir>/**/*.test.ts*'
   ],
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
@@ -210,7 +209,6 @@ module.exports = {
   },
 
   // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation
-  transformIgnorePatterns: [`/node_modules/(?!${esModulesToExclude})`],
   // transformIgnorePatterns: [
   //   "/!node_modules\\/lodash-es/",
   // ],