+++ CHANGELOG.md
@@ -38,6 +38,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2324 Hide certain columns when private practice is selected
 - YLP-2348 New columns from glycemia indicators in patients list
 - YLP-2350 Display a visual info when a patient has unread messages from another team in the chat widget
+- YLP-2354 Access to the DBL settings page from every view
 - YLP-2358 As a care team member, I want a new design when listing my pending patients
 - YLP-2381 As a HCP, I want to manage patient invites made by other members of the teams
 - YLP-2388 Select the new care team after creating it
@@ -49,6 +50,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2399 Cancel invite dialog new design
 - YLP-2430 Add checkbox to confirm that user is a hcp
 - YLP-2435 Remove "Telemonitored" filter on patient list
+- YLP-2447 Device page new design
 
 ### Engineering use
 - YLP-2044 Remove redux settings store
@@ -115,6 +117,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2446 Patient does not have to retrieve a list of patients
 - YLP-2454 Daily view crashes on mobile
 - YLP-2458 Blip auto dependencies update
+- YLP-2468 Fix team management issues
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team